[LOGMGR-233] Move heavier handlers to from core to ext.
