+++ ext/src/main/java/org/jboss/logmanager/ext/handlers/SslTcpOutputStream.java
@@ -1,7 +1,7 @@
 /*
  * JBoss, Home of Professional Open Source.
  *
- * Copyright 2014 Red Hat, Inc., and individual contributors
+ * Copyright 2018 Red Hat, Inc., and individual contributors
  * as indicated by the @author tags.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -17,16 +17,18 @@
  * limitations under the License.
  */
 
-package org.jboss.logmanager.handlers;
+package org.jboss.logmanager.ext.handlers;
 
 import java.io.IOException;
 import java.net.InetAddress;
 import javax.net.SocketFactory;
 import javax.net.ssl.SSLSocketFactory;
 
+import org.jboss.logmanager.handlers.FlushableCloseable;
+
 /**
  * An output stream that writes data to a {@link java.net.Socket socket}. Uses {@link
- * javax.net.ssl.SSLSocketFactory#getDefault()} to create the socket.
+ * SSLSocketFactory#getDefault()} to create the socket.
  *
  * @author <a href="mailto:jperkins@redhat.com">James R. Perkins</a>
  */
@@ -36,7 +38,7 @@ public class SslTcpOutputStream extends TcpOutputStream implements FlushableClos
     /**
      * Creates a SSL TCP output stream.
      * <p/>
-     * Uses the {@link javax.net.ssl.SSLSocketFactory#getDefault() default socket factory} to create the socket.
+     * Uses the {@link SSLSocketFactory#getDefault() default socket factory} to create the socket.
      *
      * @param address the address to connect to
      * @param port    the port to connect to
@@ -50,7 +52,7 @@ public class SslTcpOutputStream extends TcpOutputStream implements FlushableClos
     /**
      * Creates a SSL TCP output stream.
      * <p/>
-     * Uses the {@link javax.net.ssl.SSLSocketFactory#getDefault() default socket factory} to create the socket.
+     * Uses the {@link SSLSocketFactory#getDefault() default socket factory} to create the socket.
      *
      * @param socketFactory the factory used to create the socket
      * @param address       the address to connect to
@@ -65,7 +67,7 @@ public class SslTcpOutputStream extends TcpOutputStream implements FlushableClos
     /**
      * Creates a SSL TCP output stream.
      * <p/>
-     * Uses the {@link javax.net.ssl.SSLSocketFactory#getDefault() default socket factory} to create the socket.
+     * Uses the {@link SSLSocketFactory#getDefault() default socket factory} to create the socket.
      *
      * @param address          the address to connect to
      * @param port             the port to connect to
@@ -81,7 +83,7 @@ public class SslTcpOutputStream extends TcpOutputStream implements FlushableClos
     /**
      * Creates a SSL TCP output stream.
      * <p/>
-     * Uses the {@link javax.net.ssl.SSLSocketFactory#getDefault() default socket factory} to create the socket.
+     * Uses the {@link SSLSocketFactory#getDefault() default socket factory} to create the socket.
      *
      * @param socketFactory    the factory used to create the socket
      * @param address          the address to connect to
