+++ ext/src/main/java/org/jboss/logmanager/ext/handlers/SuffixRotator.java
@@ -1,7 +1,7 @@
 /*
  * JBoss, Home of Professional Open Source.
  *
- * Copyright 2017 Red Hat, Inc., and individual contributors
+ * Copyright 2018 Red Hat, Inc., and individual contributors
  * as indicated by the @author tags.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -17,7 +17,7 @@
  * limitations under the License.
  */
 
-package org.jboss.logmanager.handlers;
+package org.jboss.logmanager.ext.handlers;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -109,7 +109,7 @@ class SuffixRotator {
     }
 
     /**
-     * The {@linkplain java.text.SimpleDateFormat date format pattern} for the suffix or an empty
+     * The {@linkplain SimpleDateFormat date format pattern} for the suffix or an empty
      * {@linkplain String string}.
      *
      * @return the date pattern or an empty string
