+++ ext/src/test/java/org/jboss/logmanager/ext/handlers/AbstractHandlerTest.java
@@ -1,7 +1,7 @@
 /*
  * JBoss, Home of Professional Open Source.
  *
- * Copyright 2014 Red Hat, Inc., and individual contributors
+ * Copyright 2018 Red Hat, Inc., and individual contributors
  * as indicated by the @author tags.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -17,7 +17,7 @@
  * limitations under the License.
  */
 
-package org.jboss.logmanager.handlers;
+package org.jboss.logmanager.ext.handlers;
 
 import java.io.BufferedReader;
 import java.io.File;
@@ -47,7 +47,7 @@ public class AbstractHandlerTest {
     static final File BASE_LOG_DIR;
 
     static {
-        BASE_LOG_DIR = new File(System.getProperty("test.log.dir"));
+        BASE_LOG_DIR = new File(System.getProperty("log.dir"));
     }
 
     final static PatternFormatter FORMATTER = new PatternFormatter("%d{HH:mm:ss,SSS} %-5p [%c] (%t) %s%E%n");
