+++ ext/src/test/resources/client.cer
