+++ ext/src/test/resources/generate.sh
