+++ ext/src/test/resources/test-server-store.jks
